refactor: 모델 세부 패키지 분리
