+++ src/main/java/controller/LadderController.java
@@ -4,11 +4,11 @@ import dto.ParticipantName;
 import java.util.List;
 import java.util.function.Function;
 import java.util.function.Supplier;
-import model.GameResults;
-import model.Height;
-import model.Ladder;
-import model.Participants;
-import model.Prizes;
+import model.ladder.Height;
+import model.ladder.Ladder;
+import model.participant.Participants;
+import model.prize.Prizes;
+import model.result.GameResults;
 import view.InputView;
 import view.OutputView;
 import view.ResultView;
