+++ src/main/java/dto/LayerSteps.java
@@ -1,8 +1,8 @@
 package dto;
 
 import java.util.List;
-import model.Layer;
-import model.Step;
+import model.ladder.Layer;
+import model.ladder.Step;
 
 public record LayerSteps(List<Step> steps) {
     public LayerSteps(Layer layer) {
