+++ src/main/java/model/generator/StepExistenceGenerator.java
@@ -1,8 +1,9 @@
-package model;
+package model.generator;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
+import model.ladder.Step;
 
 public class StepExistenceGenerator {
     private static final Random random = new Random();
