+++ src/main/java/model/prize/Prize.java
@@ -1,4 +1,4 @@
-package model;
+package model.prize;
 
 import dto.PrizeName;
 import utils.Constant;
