+++ src/main/java/model/prize/Prizes.java
@@ -1,4 +1,4 @@
-package model;
+package model.prize;
 
 import dto.PrizeName;
 import java.util.List;
