+++ src/test/java/model/LadderTest.java
@@ -4,6 +4,13 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 import java.util.Map;
+import model.ladder.Ladder;
+import model.ladder.Layer;
+import model.ladder.Step;
+import model.participant.Participant;
+import model.participant.Participants;
+import model.prize.Prize;
+import model.prize.Prizes;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
