+++ src/test/java/model/ParticipantsTest.java
@@ -3,6 +3,7 @@ package model;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
+import model.participant.Participants;
 import org.junit.jupiter.api.Test;
 
 public class ParticipantsTest {
