+++ src/test/java/model/PrizeTest.java
@@ -2,6 +2,7 @@ package model;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import model.prize.Prize;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
