+++ src/test/java/model/PrizesTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
+import model.prize.Prizes;
 import org.junit.jupiter.api.Test;
 
 public class PrizesTest {
